refactor: 서버 재요청 로직 hook 분리
